requirements.txt
asgiref==3.5.0
autopep8==1.6.0
btrfsutil==5.16
CacheControl==0.12.6
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==36.0.1
distlib==0.3.4
distro==1.6.0
dj-database-url==0.5.0
Django==4.0.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
dnspython==2.2.0
filelock==3.4.0
Flask==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.10.0
msgpack==1.0.3
ordered-set==4.0.2
packaging==20.9
pep517==0.12.0
Pillow==9.0.1
platformdirs==2.4.1
ply==3.11
progress==1.6
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pylint==2.12.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytz==2021.3
Reflector==2021.11.20.2.41.3
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
six==1.16.0
smbus==1.1
sqlparse==0.4.2
team==1.0
toml==0.10.2
tomli==2.0.0
typed-ast==1.5.2
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.11.0
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.7.0
